requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
psycopg2-binary
pandas
numpy
python-dotenv
pydantic
sqlalchemy
asyncpg
python-multipart
pypdf
tabula-py
camelot-py[cv]
pdfplumber
openpyxl
PyPDF2
pdfminer.six
google-generativeai



